requirements.txt
flask
flask-socketio
eventlet
psycopg2-binary
pyserial
